pyproject.toml
[project]
name = "ecommerce-analytics"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2.0",
  "duckdb>=1.1.0",
  "pyarrow>=16.0.0",
  "kaggle>=1.6.0",
  "python-dateutil>=2.9.0",
  "google-cloud-bigquery>=3.20.0",
  "pandas-gbq>=0.22.0",
]

